+++ lib/src/main/kotlin/com/example/HelloWorld.kt
@@ -1,0 +1,7 @@
+package com.example
+
+class HelloWorld {
+    fun someLibraryMethod(): Boolean {
+        return true
+    }
+}
