--- lib/src/main/kotlin/hello/world/Library.kt
@@ -1,10 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package hello.world
-
-class Library {
-    fun someLibraryMethod(): Boolean {
-        return true
-    }
-}
